+++ src/parse.ml
@@ -59,6 +59,7 @@ and parse_expression sexp =
   match sexp with
   | Sym id -> if is_primitive id then Keyword id else Identifier id
   | Lit lit -> Literal lit
+  | List (Sym "tuple" :: bindings) -> TupleExpression (List.map parse_tuple_binding bindings)
   | List [Sym "some"; expr] -> SomeExpression (parse_expression expr)
   | List (Sym "list" :: exprs) -> ListExpression (List.map parse_expression exprs)
   | List [Sym "is-none"; expr] -> IsNone (parse_expression expr)
@@ -96,6 +97,10 @@ and parse_expression sexp =
   | List [Sym "to-uint"; expr] -> ToUint (parse_expression expr)
   | List (Sym name :: args) -> FunctionCall (name, (List.map parse_expression args))
   | List _ -> failwith "invalid Clarity expression"
+
+and parse_tuple_binding = function
+  | List [Sym k; v] -> (k, parse_expression v)
+  | _ -> failwith "invalid Clarity tuple expression"
 
 and parse_type = function
   | Sym "principal" -> Principal