+++ CHANGES.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clarity keywords (such as `block-height`) are now represented in the AST
   with `Keyword` nodes instead of `Identifier` nodes.
 
+### Added
+
+- Support for tuple expressions of the form `(tuple (name "blockstack") (id 1337))`.
+
 ## 0.1.0 - 2020-09-30
 
 ### Added